chore: reorganize fonts and update theme configuration
- Remove old font files:
  - fonts/inter/Inter_Regular.ttf
  - fonts/roboto/Roboto-Regular.ttf
- Relocate and update font management:
  - Move fonts to assets/fonts/
  - Update font references in app_theme.dart
  - Replace lib/core/utils/app_fonts_family.dart with lib/core/assets/app_fonts_family.dart
- Update pubspec.yaml to reflect new font paths
