+++ assets/fonts/inter/Inter_Regular.ttf
