+++ lib/core/assets/app_fonts_family.dart
@@ -1,4 +1,4 @@
 class AppFontsFamily{
-  static const String intel="Inter";
+  static const String inter="Inter";
   static const String roboto="Roboto";
 }
